001-chatbot/requirements.txt
langchain
langchain-openai
langchain-community
streamlit
python-dotenv
langsmith